Move file system watching tests to one class
