+++ subprojects/file-watching/src/integTest/groovy/org/gradle/internal/watch/EnableFileSystemWatchingIntegrationTest.groovy
@@ -16,10 +16,7 @@
 
 package org.gradle.internal.watch
 
-import net.rubygrapefruit.platform.internal.jni.NativeLogger
 import org.gradle.initialization.StartParameterBuildOptions
-import org.gradle.integtests.fixtures.daemon.DaemonFixture
-import org.gradle.integtests.fixtures.daemon.DaemonLogsAnalyzer
 import org.gradle.internal.service.scopes.VirtualFileSystemServices
 import spock.lang.Unroll
 
@@ -79,79 +76,6 @@ class EnableFileSystemWatchingIntegrationTest extends AbstractFileSystemWatching
         "--no-watch-fs"   | DISABLED_MESSAGE
     }
 
-    def "verbose VFS logging can be enabled"() {
-        buildFile << """
-            apply plugin: "java"
-        """
-        when:
-        withWatchFs().run("assemble", "-D${StartParameterBuildOptions.VfsVerboseLoggingOption.GRADLE_PROPERTY}=true")
-        then:
-        !(result.output =~ /Received \d+ file system events since last build while watching \d+ hierarchies/)
-        !(result.output =~ /Virtual file system retained information about \d+ files, \d+ directories and \d+ missing files since last build/)
-        result.output =~ /Received \d+ file system events during the current build while watching \d+ hierarchies/
-        result.output =~ /Virtual file system retains information about \d+ files, \d+ directories and \d+ missing files until next build/
-
-        when:
-        withWatchFs().run("assemble", "-D${StartParameterBuildOptions.VfsVerboseLoggingOption.GRADLE_PROPERTY}=true")
-        then:
-        result.output =~ /Received \d+ file system events since last build while watching \d+ hierarchies/
-        result.output =~ /Virtual file system retained information about \d+ files, \d+ directories and \d+ missing files since last build/
-        result.output =~ /Received \d+ file system events during the current build while watching \d+ hierarchies/
-        result.output =~ /Virtual file system retains information about \d+ files, \d+ directories and \d+ missing files until next build/
-
-        when:
-        withWatchFs().run("assemble", "-D${StartParameterBuildOptions.VfsVerboseLoggingOption.GRADLE_PROPERTY}=false")
-        then:
-        !(result.output =~ /Received \d+ file system events since last build while watching \d+ hierarchies/)
-        !(result.output =~ /Virtual file system retained information about \d+ files, \d+ directories and \d+ missing files since last build/)
-        !(result.output =~ /Received \d+ file system events during the current build while watching \d+ hierarchies/)
-        !(result.output =~ /Virtual file system retains information about \d+ files, \d+ directories and \d+ missing files until next build/)
-    }
-
-    def "debug logging can be enabled"() {
-        executer.beforeExecute {
-            executer.requireDaemon()
-            executer.requireIsolatedDaemons()
-        }
-
-        buildFile << """
-            tasks.register("change") {
-                def output = file("output.txt")
-                outputs.file(output)
-                outputs.upToDateWhen { false }
-                doLast {
-                    output.text = Math.random() as String
-                }
-            }
-        """
-
-        when:
-        withWatchFs().run("change", "--debug", "-D${StartParameterBuildOptions.WatchFileSystemDebugLoggingOption.GRADLE_PROPERTY}=true")
-        waitForChangesToBePickedUp()
-        then:
-        output.contains(NativeLogger.name)
-
-        when:
-        def logLineCountBeforeChange = daemon.logLineCount
-        file("output.txt").text = "Changed"
-        waitForChangesToBePickedUp()
-        then:
-        daemon.logContains(logLineCountBeforeChange, NativeLogger.name)
-
-        when:
-        withWatchFs().run("change", "--debug", "-D${StartParameterBuildOptions.WatchFileSystemDebugLoggingOption.GRADLE_PROPERTY}=false")
-        waitForChangesToBePickedUp()
-        then:
-        !output.contains(NativeLogger.name)
-
-        when:
-        logLineCountBeforeChange = daemon.logLineCount
-        file("output.txt").text = "Changed"
-        waitForChangesToBePickedUp()
-        then:
-        !daemon.logContains(logLineCountBeforeChange, NativeLogger.name)
-    }
-
     @Unroll
     @SuppressWarnings('GrDeprecatedAPIUsage')
     def "deprecation message is shown when using the old property to enable watching the file system (enabled: #enabled)"() {
@@ -206,9 +130,5 @@ class EnableFileSystemWatchingIntegrationTest extends AbstractFileSystemWatching
 
         where:
         drop << [true, false]
-    }
-
-    private DaemonFixture getDaemon() {
-        new DaemonLogsAnalyzer(executer.daemonBaseDir).daemon
     }
 }
